+++ R/gg_function.R
@@ -93,7 +93,7 @@ gg_function <- function(
   col_palette = NULL,
   col_na = NULL,
 
-  col_rescale = scales::rescale(),
+  col_rescaler = scales::rescale,
   col_scale_type = "gradient",
   col_transform = NULL,
   colour_palette = NULL,
@@ -194,7 +194,7 @@ gg_function <- function(
     col_palette = col_palette,
     col_na = col_na,
 
-    col_rescale = col_rescale,
+    col_rescaler = col_rescaler,
     col_scale_type = col_scale_type,
     col_title = col_title,
     col_transform = col_transform,